+++ src/formatDistance/test.js
@@ -274,19 +274,25 @@ describe('formatDistance', function() {
     })
   })
 
-  it("returns String('Invalid Date') if the first date is `Invalid Date`", function() {
-    var result = formatDistance(new Date(NaN), new Date(1986, 3, 7, 10, 32, 0))
-    assert(result === 'Invalid Date')
+  it('throws RangeError if the first date is `Invalid Date`', function() {
+    assert.throws(
+      formatDistance.bind(null, new Date(NaN), new Date(1986, 3, 7, 10, 32, 0)),
+      RangeError
+    )
   })
 
-  it("returns String('Invalid Date') if the second date is `Invalid Date`", function() {
-    var result = formatDistance(new Date(1986, 3, 4, 10, 32, 0), new Date(NaN))
-    assert(result === 'Invalid Date')
+  it('throws RangeError if the second date is `Invalid Date`', function() {
+    assert.throws(
+      formatDistance.bind(null, new Date(1986, 3, 4, 10, 32, 0), new Date(NaN)),
+      RangeError
+    )
   })
 
-  it("returns String('Invalid Date') if the both dates are `Invalid Date`", function() {
-    var result = formatDistance(new Date(NaN), new Date(NaN))
-    assert(result === 'Invalid Date')
+  it('throws RangeError if the both dates are `Invalid Date`', function() {
+    assert.throws(
+      formatDistance.bind(null, new Date(NaN), new Date(NaN)),
+      RangeError
+    )
   })
 
   it('throws TypeError exception if passed less than 2 arguments', function() {