+++ src/formatDistanceStrict/index.js
@@ -178,7 +178,7 @@ export default function formatDistanceStrict(
   var comparison = compareAsc(dirtyDate, dirtyBaseDate)
 
   if (isNaN(comparison)) {
-    return 'Invalid Date'
+    throw new RangeError('Invalid time value')
   }
 
   var localizeOptions = cloneObject(options)