+++ src/formatDistanceStrict/test.js
@@ -442,25 +442,33 @@ describe('formatDistanceStrict', function() {
     })
   })
 
-  it("returns String('Invalid Date') if the first date is `Invalid Date`", function() {
-    var result = formatDistanceStrict(
-      new Date(NaN),
-      new Date(1986, 3, 7, 10, 32, 0)
+  it('throws `RangeError` if the first date is `Invalid Date`', function() {
+    assert.throws(
+      formatDistanceStrict.bind(
+        null,
+        new Date(NaN),
+        new Date(1986, 3, 7, 10, 32, 0)
+      ),
+      RangeError
     )
-    assert(result === 'Invalid Date')
   })
 
-  it("returns String('Invalid Date') if the second date is `Invalid Date`", function() {
-    var result = formatDistanceStrict(
-      new Date(1986, 3, 4, 10, 32, 0),
-      new Date(NaN)
+  it('throws `RangeError` if the second date is `Invalid Date`', function() {
+    assert.throws(
+      formatDistanceStrict.bind(
+        null,
+        new Date(1986, 3, 4, 10, 32, 0),
+        new Date(NaN)
+      ),
+      RangeError
     )
-    assert(result === 'Invalid Date')
   })
 
-  it("returns String('Invalid Date') if the both dates are `Invalid Date`", function() {
-    var result = formatDistanceStrict(new Date(NaN), new Date(NaN))
-    assert(result === 'Invalid Date')
+  it('throws `RangeError` if the both dates are `Invalid Date`', function() {
+    assert.throws(
+      formatDistanceStrict.bind(null, new Date(NaN), new Date(NaN)),
+      RangeError
+    )
   })
 
   it("throws `RangeError` if `options.roundingMethod` is not 'floor', 'ceil', 'round' or undefined", function() {